requirements.txt
fastapi==0.115.12
python-multipart==0.0.20
python-docx==1.1.2
uvicorn[standard]==0.34.0
